Implement Rover Rate Control
Adds Rover Attitude Control class to enable rover rate control. The relevant parameters were defined.

For rate control, a PI controller with a feedforward term was added. The steering angle is sclaled with the inverse of the groundspeed. This is based on the kinematics of a ackerman steering vehicle.

Differential rovers are not handled yet